add post – fassi.live
oh hey there, you're still here. good. this project in particular has been ... a whilwind. it went from idea in my head to mvp and then 3-part series and accompanying post and so on in ... 24h? i don't know. is this the power of tje human-assisted ai we live in? offload boring tasks to ai, spend time in the "reality building" and creative space, and enjoy life? it's not too bad if you ask me. but at the same time, this is how every end of the world prophecy starts – "it's not too bad".